+++ code/kurma-server/build.sh
@@ -11,8 +11,10 @@ if [[ -f $BASE_PATH/version/number ]]; then
 fi
 BUILD_LDFLAGS="-X github.com/apcera/kurma/stage1/client.version=$version"
 
-mkdir -p go/src/github.com/apcera
-ln -s $BASE_PATH/kurma-source go/src/github.com/apcera/kurma
+if [[ ! -f go/src/github.com/apcera/kurma ]]; then
+    mkdir -p go/src/github.com/apcera
+    ln -s $BASE_PATH/kurma-source go/src/github.com/apcera/kurma
+fi
 export GOPATH="$BASE_PATH/go"
 TARGET=$BASE_PATH/go/src/github.com/apcera/kurma
